commit backend

+++ src/main/java/com/example/deporsm/dto/VecinoDTO.java
@@ -1,14 +1,13 @@
 package com.example.deporsm.dto;
 
 public class VecinoDTO {
-    private Integer id;
+    private Long id;
     private String nombre;
     private String email;
     private String telefono;
-    private Integer reservas;  // Número de reservas realizadas por el vecino
-
-    // Constructor
-    public VecinoDTO(Integer id, String nombre, String email, String telefono, Integer reservas) {
+    private Long reservas;  // Número de reservas realizadas por el vecino    // Constructor que debe coincidir EXACTAMENTE con el orden de la query
+    
+    public VecinoDTO(Long id, String nombre, String email, String telefono, Long reservas) {
         this.id = id;
         this.nombre = nombre;
         this.email = email;
@@ -17,15 +16,15 @@ public class VecinoDTO {
     }
 
     // Getters y Setters
-    public Integer getId() { return id; }
+    public Long getId() { return id; }
     public String getNombre() { return nombre; }
     public String getEmail() { return email; }
     public String getTelefono() { return telefono; }
-    public Integer getReservas() { return reservas; }
+    public Long getReservas() { return reservas; }
 
-    public void setId(Integer id) { this.id = id; }
+    public void setId(Long id) { this.id = id; }
     public void setNombre(String nombre) { this.nombre = nombre; }
     public void setEmail(String email) { this.email = email; }
     public void setTelefono(String telefono) { this.telefono = telefono; }
-    public void setReservas(Integer reservas) { this.reservas = reservas; }
+    public void setReservas(Long reservas) { this.reservas = reservas; }
 }
